refactor: calcCart 함수 내 기능 함수 분리
