+++ src/basic/utils/const.js
@@ -1,0 +1,21 @@
+export const PROD_LIST = [
+  { id: 'p1', name: '상품1', price: 10000, quantity: 50 },
+  { id: 'p2', name: '상품2', price: 20000, quantity: 30 },
+  { id: 'p3', name: '상품3', price: 30000, quantity: 20 },
+  { id: 'p4', name: '상품4', price: 15000, quantity: 0 },
+  { id: 'p5', name: '상품5', price: 25000, quantity: 10 },
+];
+
+export const PRODUCT_BULK_DISCOUNT_AMOUNT = 10;
+export const PRODUCT_BULK_DISCOUNT_RATE = {
+  p1: 0.1,
+  p2: 0.15,
+  p3: 0.2,
+  p4: 0.05,
+  p5: 0.25,
+};
+
+export const TOTAL_BULK_DISCOUNT_AMOUNT = 30;
+
+export const SALE_DAY = 2;
+export const SALE_DAY_DISCOUNT_RATE = 0.1;
